pyproject.toml
[project]
name = "docs-navigator-mcp"
version = "0.1.0"
description = "Docs Navigator MCP — FastMCP server + agent + Gradio UI"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "mcp[cli]>=0.1.0",
    "anthropic>=0.36.0",
    "python-dotenv>=1.0.1",
    "gradio>=5.0.0",
]

[tool.uv]
# optional, good defaults
package = true
